Added user management,
Added a login page that allow user to login with a user name, this user name is set on the tasks created by the users.
Added a new page for displaying you tasks, will list all the task created by the current user.
Added some styling to make the status look nicer by using badges and a string instead of the int.